refactor(traces-v2): address all deferred CodeRabbit review items
Cascades + structural splits the previous review round flagged but the
in-flight comment loop punted to a follow-up. Now landing them here.

Named-params refactors (project rule: fn({ a, b }) not fn(a, b)):
- useFilterSidebarData wrappers: toggleFacet ({ field, value, isModifierKey }),
  setRange ({ field, from, to }), removeRange ({ field })
- SectionRenderer + EvaluatorDrilldown prop signatures + every call site
- FilterSidebar.renderSection ({ key, dragHandleProps })
- discoverCache.setCachedDiscover ({ projectId, facets })
- getVercelAIModel ({ projectId, model, featureKey }) — and 10+ call sites
  across ai-query, workflows, translate, dataset-generate, scenario-generate,
  simulation-runner, prompt-config adapter, plus the unit tests

Boolean naming (project rule: is/has/should/can/will prefix):
- spansLoading → isSpansLoading on SpanDetailPane, PaneLayout,
  TraceAccordions, TraceDrawerShell (4 files)
- modifierKey option on the sidebar's toggleFacet wrapper → isModifierKey

EvaluatorDrilldown split:
- Main exported function down from ~175 lines to ~65 by extracting
  VerdictSection, VerdictPill, ScoreSection, SectionLabel as local
  components + buildVerdictSpecs / computeActiveVerdicts helpers

FlameView split (1075 → 516 lines):
- New hooks: useFlameViewport, useFlamePanDrag, useFlameAxisZoom,
  useFlameKeyboard
- New components: FlameAxis, FlameBreadcrumbs, FlameContextStrip
- FlameView.tsx retains the orchestrator: state, derivations, virtualizer,
  Minimap mount

Typecheck: pre-existing errors (Box as="button"/"input" type=, ai-query
optional-chaining, api.tsx modelProviders) remain unchanged. No new
typecheck errors introduced.
