+++ langwatch/src/features/traces-v2/components/TraceDrawer/flameView/FlameView.tsx
@@ -1,28 +1,23 @@
-import { Box, Flex, HStack, Icon, Text } from "@chakra-ui/react";
+import { Box, Flex, Text } from "@chakra-ui/react";
 import { useVirtualizer } from "@tanstack/react-virtual";
 import { memo, useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { LuChevronRight, LuRotateCcw } from "react-icons/lu";
-import { Kbd } from "~/components/ops/shared/Kbd";
-import { Tooltip } from "~/components/ui/tooltip";
 import { formatDuration } from "../../../utils/formatters";
 import {
-  DRAG_THRESHOLD_PX,
-  MIN_VIEWPORT_MS,
   ROW_GAP,
   ROW_HEIGHT,
-  WHEEL_ZOOM_SENSITIVITY,
-  ZOOM_ANIMATION_MS,
   ZOOM_FIT_PADDING,
 } from "./constants";
+import { FlameAxis } from "./FlameAxis";
+import { FlameBreadcrumbs } from "./FlameBreadcrumbs";
+import { FlameContextStrip } from "./FlameContextStrip";
 import { FlameRow } from "./FlameRow";
 import { Minimap } from "./Minimap";
-import {
-  buildTree,
-  computeSpanContext,
-  formatPercent,
-  generateTicks,
-} from "./tree";
+import { buildTree, computeSpanContext, generateTicks } from "./tree";
 import type { FlameNode, FlameViewProps, SpanContext, Viewport } from "./types";
+import { useFlameAxisZoom } from "./useFlameAxisZoom";
+import { useFlamePanDrag } from "./useFlamePanDrag";
+import { useFlameKeyboard } from "./useFlameKeyboard";
+import { useFlameViewport } from "./useFlameViewport";
 
 export const FlameView = memo(function FlameView({
   spans,
@@ -43,200 +38,34 @@ export const FlameView = memo(function FlameView({
     return { startMs: start, endMs: end };
   }, [spans]);
 
-  const [viewport, setViewport] = useState<Viewport>(fullRange);
   const [hoveredSpanId, setHoveredSpanId] = useState<string | null>(null);
   const [focusedSpanId, setFocusedSpanId] = useState<string | null>(null);
-  const [dragSelection, setDragSelection] = useState<Viewport | null>(null);
 
   const containerRef = useRef<HTMLDivElement>(null);
   const flameAreaRef = useRef<HTMLDivElement>(null);
   const timeAxisRef = useRef<HTMLDivElement>(null);
-  const animationRef = useRef<number | null>(null);
-  const isPanningRef = useRef(false);
-  const viewportRef = useRef(viewport);
-  viewportRef.current = viewport;
 
-  // Reset viewport when underlying spans change.
-  useEffect(() => {
-    setViewport(fullRange);
-  }, [fullRange]);
+  const { viewport, setViewport, viewportRef, clampViewport, animateTo, cancelAnimation } =
+    useFlameViewport({ fullRange, flameAreaRef });
 
-  const cancelAnimation = useCallback(() => {
-    if (animationRef.current !== null) {
-      cancelAnimationFrame(animationRef.current);
-      animationRef.current = null;
-    }
-  }, []);
+  const { isPanningRef, handlePointerDown } = useFlamePanDrag({
+    flameAreaRef,
+    viewportRef,
+    cancelAnimation,
+    clampViewport,
+    setViewport,
+  });
 
-  // Drawer can close mid-animation; without this, the rAF tick keeps firing
-  // setViewport on an unmounted component.
-  useEffect(() => () => cancelAnimation(), [cancelAnimation]);
+  const { dragSelection, handleTimeAxisPointerDown } = useFlameAxisZoom({
+    timeAxisRef,
+    viewportRef,
+    cancelAnimation,
+    animateTo,
+  });
 
-  const clampViewport = useCallback(
-    (v: Viewport): Viewport => {
-      const fullDur = fullRange.endMs - fullRange.startMs;
-      if (fullDur <= 0) return fullRange;
-      const minDur = Math.min(MIN_VIEWPORT_MS, fullDur);
-      const dur = Math.max(minDur, Math.min(fullDur, v.endMs - v.startMs));
-      let start = v.startMs;
-      let end = start + dur;
-      if (start < fullRange.startMs) {
-        start = fullRange.startMs;
-        end = start + dur;
-      }
-      if (end > fullRange.endMs) {
-        end = fullRange.endMs;
-        start = end - dur;
-      }
-      return { startMs: start, endMs: end };
-    },
-    [fullRange],
-  );
-
-  const animateTo = useCallback(
-    (target: Viewport) => {
-      cancelAnimation();
-      const clamped = clampViewport(target);
-      const from = viewportRef.current;
-      const startTime = performance.now();
-      const tick = (now: number) => {
-        const t = Math.min(1, (now - startTime) / ZOOM_ANIMATION_MS);
-        const e = 1 - Math.pow(1 - t, 3);
-        setViewport({
-          startMs: from.startMs + (clamped.startMs - from.startMs) * e,
-          endMs: from.endMs + (clamped.endMs - from.endMs) * e,
-        });
-        if (t < 1) {
-          animationRef.current = requestAnimationFrame(tick);
-        } else {
-          animationRef.current = null;
-        }
-      };
-      animationRef.current = requestAnimationFrame(tick);
-    },
-    [cancelAnimation, clampViewport],
-  );
-
-  // Selection-follow: when a span is selected externally and falls fully outside
-  // the current viewport, animate the viewport to bring it back into view.
-  useEffect(() => {
-    if (!selectedSpanId) return;
-    const node = tree.byId.get(selectedSpanId);
-    if (!node) return;
-    const v = viewportRef.current;
-    const isCompletelyOutside =
-      node.span.endTimeMs < v.startMs || node.span.startTimeMs > v.endMs;
-    if (!isCompletelyOutside) return;
-    const dur = node.span.endTimeMs - node.span.startTimeMs;
-    const vpDur = v.endMs - v.startMs;
-    if (dur < vpDur * 0.5) {
-      // Span is small relative to current zoom — keep zoom level, just center it.
-      const center = (node.span.startTimeMs + node.span.endTimeMs) / 2;
-      animateTo({
-        startMs: center - vpDur / 2,
-        endMs: center + vpDur / 2,
-      });
-    } else {
-      const pad = Math.max(dur * ZOOM_FIT_PADDING, 0);
-      animateTo({
-        startMs: node.span.startTimeMs - pad,
-        endMs: node.span.endTimeMs + pad,
-      });
-    }
-  }, [selectedSpanId, tree.byId, animateTo]);
-
-  // Wheel: zoom toward cursor (deltaY) or pan (deltaX / shift).
-  useEffect(() => {
-    const el = flameAreaRef.current;
-    if (!el) return;
-    const handleWheel = (e: WheelEvent) => {
-      e.preventDefault();
-      cancelAnimation();
-      const rect = el.getBoundingClientRect();
-      const x = Math.max(0, Math.min(1, (e.clientX - rect.left) / rect.width));
-      const isPan = e.shiftKey || Math.abs(e.deltaX) > Math.abs(e.deltaY);
-      const delta = isPan ? e.deltaX || e.deltaY : e.deltaY;
-      setViewport((prev) => {
-        const dur = prev.endMs - prev.startMs;
-        if (isPan) {
-          const dt = (delta / rect.width) * dur;
-          return clampViewport({
-            startMs: prev.startMs + dt,
-            endMs: prev.endMs + dt,
-          });
-        }
-        const cursorTime = prev.startMs + x * dur;
-        const factor = Math.exp(delta * WHEEL_ZOOM_SENSITIVITY);
-        const newDur = dur * factor;
-        const newStart = cursorTime - x * newDur;
-        return clampViewport({
-          startMs: newStart,
-          endMs: newStart + newDur,
-        });
-      });
-    };
-    el.addEventListener("wheel", handleWheel, { passive: false });
-    return () => el.removeEventListener("wheel", handleWheel);
-  }, [cancelAnimation, clampViewport]);
-
-  // Drag-to-pan on the flame area; spans get click events on no-drag.
-  const handlePointerDown = useCallback(
-    (e: React.PointerEvent) => {
-      if (e.button !== 0) return;
-      const el = flameAreaRef.current;
-      if (!el) return;
-      const rect = el.getBoundingClientRect();
-      const startX = e.clientX;
-      const startVp = viewportRef.current;
-      const dur = startVp.endMs - startVp.startMs;
-      let dragged = false;
-      cancelAnimation();
-
-      const handleMove = (ev: PointerEvent) => {
-        const dx = ev.clientX - startX;
-        if (!dragged && Math.abs(dx) >= DRAG_THRESHOLD_PX) {
-          dragged = true;
-          isPanningRef.current = true;
-          document.body.style.cursor = "grabbing";
-        }
-        if (!dragged) return;
-        const dt = (dx / rect.width) * dur;
-        setViewport(
-          clampViewport({
-            startMs: startVp.startMs - dt,
-            endMs: startVp.endMs - dt,
-          }),
-        );
-      };
-
-      const cleanup = () => {
-        window.removeEventListener("pointermove", handleMove);
-        window.removeEventListener("pointerup", cleanup);
-        window.removeEventListener("pointercancel", cleanup);
-        window.removeEventListener("blur", cleanup);
-        document.body.style.cursor = "";
-        // Defer flag reset so synchronous click handlers see we just dragged.
-        setTimeout(() => {
-          isPanningRef.current = false;
-        }, 0);
-      };
-
-      window.addEventListener("pointermove", handleMove);
-      window.addEventListener("pointerup", cleanup);
-      window.addEventListener("pointercancel", cleanup);
-      window.addEventListener("blur", cleanup);
-    },
-    [cancelAnimation, clampViewport],
-  );
-
-  const handleSpanClick = useCallback(
-    (spanId: string) => {
-      if (isPanningRef.current) return;
-      onSelectSpan(spanId);
-      setFocusedSpanId(spanId);
-    },
-    [onSelectSpan],
-  );
+  const handleResetZoom = useCallback(() => {
+    animateTo(fullRange);
+  }, [animateTo, fullRange]);
 
   const handleSpanDoubleClick = useCallback(
     (spanId: string) => {
@@ -254,9 +83,14 @@ export const FlameView = memo(function FlameView({
     [tree.byId, animateTo, onSelectSpan],
   );
 
-  const handleResetZoom = useCallback(() => {
-    animateTo(fullRange);
-  }, [animateTo, fullRange]);
+  const handleSpanClick = useCallback(
+    (spanId: string) => {
+      if (isPanningRef.current) return;
+      onSelectSpan(spanId);
+      setFocusedSpanId(spanId);
+    },
+    [onSelectSpan, isPanningRef],
+  );
 
   const handleClearOnEmpty = useCallback(
     (e: React.MouseEvent) => {
@@ -270,8 +104,56 @@ export const FlameView = memo(function FlameView({
       }
       onClearSpan();
     },
-    [onClearSpan],
+    [onClearSpan, isPanningRef],
   );
+
+  const dur = viewport.endMs - viewport.startMs;
+  const fullDur = fullRange.endMs - fullRange.startMs;
+  const isZoomed = fullDur > 0 && dur < fullDur * 0.999;
+
+  useFlameKeyboard({
+    containerRef,
+    tree,
+    fullDur,
+    selectedSpanId,
+    focusedSpanId,
+    setFocusedSpanId,
+    viewportRef,
+    setViewport,
+    clampViewport,
+    handleResetZoom,
+    handleSpanDoubleClick,
+    onClearSpan,
+    onSelectSpan,
+  });
+
+  // Selection-follow: when a span is selected externally and falls fully outside
+  // the current viewport, animate the viewport to bring it back into view.
+  useEffect(() => {
+    if (!selectedSpanId) return;
+    const node = tree.byId.get(selectedSpanId);
+    if (!node) return;
+    const v = viewportRef.current;
+    const isCompletelyOutside =
+      node.span.endTimeMs < v.startMs || node.span.startTimeMs > v.endMs;
+    if (!isCompletelyOutside) return;
+    const nodeDur = node.span.endTimeMs - node.span.startTimeMs;
+    const vpDur = v.endMs - v.startMs;
+    if (nodeDur < vpDur * 0.5) {
+      // Span is small relative to current zoom — keep zoom level, just center it.
+      const center = (node.span.startTimeMs + node.span.endTimeMs) / 2;
+      animateTo({
+        startMs: center - vpDur / 2,
+        endMs: center + vpDur / 2,
+      });
+    } else {
+      const pad = Math.max(nodeDur * ZOOM_FIT_PADDING, 0);
+      animateTo({
+        startMs: node.span.startTimeMs - pad,
+        endMs: node.span.endTimeMs + pad,
+      });
+    }
+  }, [selectedSpanId, tree.byId, animateTo, viewportRef]);
 
   // Ancestor chain of the focus span for breadcrumb navigation.
   const breadcrumbs = useMemo(() => {
@@ -331,68 +213,6 @@ export const FlameView = memo(function FlameView({
     };
   }, [contextNode]);
 
-  // Drag-to-zoom on the time axis: drag horizontally to select a range, release to animate-zoom.
-  const handleTimeAxisPointerDown = useCallback(
-    (e: React.PointerEvent) => {
-      if (e.button !== 0) return;
-      const el = timeAxisRef.current;
-      if (!el) return;
-      e.preventDefault();
-      cancelAnimation();
-      const rect = el.getBoundingClientRect();
-      const startVp = viewportRef.current;
-      const startDur = startVp.endMs - startVp.startMs;
-      const xToTime = (clientX: number) => {
-        const x = Math.max(0, Math.min(1, (clientX - rect.left) / rect.width));
-        return startVp.startMs + x * startDur;
-      };
-      const startTimeMs = xToTime(e.clientX);
-      const startClientX = e.clientX;
-      let dragged = false;
-
-      const handleMove = (ev: PointerEvent) => {
-        const dx = Math.abs(ev.clientX - startClientX);
-        if (!dragged && dx >= DRAG_THRESHOLD_PX) dragged = true;
-        if (!dragged) return;
-        const t = xToTime(ev.clientX);
-        setDragSelection({
-          startMs: Math.min(startTimeMs, t),
-          endMs: Math.max(startTimeMs, t),
-        });
-      };
-
-      const handleUp = (ev: PointerEvent) => {
-        cleanup();
-        if (!dragged) return;
-        const t = xToTime(ev.clientX);
-        const sel: Viewport = {
-          startMs: Math.min(startTimeMs, t),
-          endMs: Math.max(startTimeMs, t),
-        };
-        if (sel.endMs - sel.startMs >= MIN_VIEWPORT_MS) {
-          animateTo(sel);
-        }
-      };
-      const cleanup = () => {
-        window.removeEventListener("pointermove", handleMove);
-        window.removeEventListener("pointerup", handleUp);
-        window.removeEventListener("pointercancel", cleanup);
-        window.removeEventListener("blur", cleanup);
-        setDragSelection(null);
-      };
-
-      window.addEventListener("pointermove", handleMove);
-      window.addEventListener("pointerup", handleUp);
-      window.addEventListener("pointercancel", cleanup);
-      window.addEventListener("blur", cleanup);
-    },
-    [animateTo, cancelAnimation],
-  );
-
-  const dur = viewport.endMs - viewport.startMs;
-  const fullDur = fullRange.endMs - fullRange.startMs;
-  const isZoomed = fullDur > 0 && dur < fullDur * 0.999;
-
   const visibleBlocks = useMemo(() => {
     if (dur <= 0) return tree.all;
     return tree.all.filter(
@@ -448,149 +268,6 @@ export const FlameView = memo(function FlameView({
 
   const virtualRows = virtualizer.getVirtualItems();
 
-  // Keyboard navigation. Uses functional setState so the listener is stable.
-  useEffect(() => {
-    const el = containerRef.current;
-    if (!el) return;
-    const handleKeyDown = (e: KeyboardEvent) => {
-      const target = e.target as HTMLElement;
-      if (
-        target.tagName === "INPUT" ||
-        target.tagName === "TEXTAREA" ||
-        target.isContentEditable
-      )
-        return;
-      if (!el.contains(target) && target !== el) return;
-
-      switch (e.key) {
-        case "Escape": {
-          if (
-            viewportRef.current.endMs - viewportRef.current.startMs <
-            fullDur * 0.999
-          ) {
-            e.preventDefault();
-            handleResetZoom();
-          } else if (selectedSpanId) {
-            e.preventDefault();
-            onClearSpan();
-          }
-          break;
-        }
-        case "0":
-        case "Home": {
-          e.preventDefault();
-          handleResetZoom();
-          break;
-        }
-        case "Enter": {
-          if (focusedSpanId) {
-            e.preventDefault();
-            handleSpanDoubleClick(focusedSpanId);
-          }
-          break;
-        }
-        case " ": {
-          if (focusedSpanId) {
-            e.preventDefault();
-            onSelectSpan(focusedSpanId);
-          }
-          break;
-        }
-        case "ArrowLeft":
-        case "ArrowRight": {
-          // Only intercept Arrow keys when the user has actually engaged
-          // with the flame — either holding shift to pan the viewport, or
-          // navigating between sibling spans after focusing one. Without
-          // a focused span and no modifier, the drawer-level handler is
-          // the right consumer (prev/next trace in the conversation), and
-          // the previous unconditional `preventDefault()` here used to fight
-          // it depending on what had document focus.
-          const direction = e.key === "ArrowLeft" ? -1 : 1;
-          if (e.shiftKey) {
-            e.preventDefault();
-            e.stopImmediatePropagation();
-            setViewport((v) => {
-              const d = v.endMs - v.startMs;
-              const pan = d * 0.2 * direction;
-              return clampViewport({
-                startMs: v.startMs + pan,
-                endMs: v.endMs + pan,
-              });
-            });
-          } else if (focusedSpanId) {
-            const node = tree.byId.get(focusedSpanId);
-            if (node) {
-              const siblings = node.parent ? node.parent.children : tree.roots;
-              const idx = siblings.findIndex(
-                (n) => n.span.spanId === focusedSpanId,
-              );
-              const next = siblings[idx + direction];
-              if (next) {
-                e.preventDefault();
-                e.stopImmediatePropagation();
-                setFocusedSpanId(next.span.spanId);
-              }
-            }
-          }
-          break;
-        }
-        case "ArrowUp":
-        case "ArrowDown": {
-          if (!focusedSpanId) break;
-          const node = tree.byId.get(focusedSpanId);
-          if (!node) break;
-          if (e.key === "ArrowUp" && node.parent) {
-            e.preventDefault();
-            setFocusedSpanId(node.parent.span.spanId);
-          } else if (e.key === "ArrowDown" && node.children.length > 0) {
-            e.preventDefault();
-            setFocusedSpanId(node.children[0]!.span.spanId);
-          }
-          break;
-        }
-        case "+":
-        case "=": {
-          e.preventDefault();
-          setViewport((v) => {
-            const center = (v.startMs + v.endMs) / 2;
-            const newDur = (v.endMs - v.startMs) * 0.7;
-            return clampViewport({
-              startMs: center - newDur / 2,
-              endMs: center + newDur / 2,
-            });
-          });
-          break;
-        }
-        case "-":
-        case "_": {
-          e.preventDefault();
-          setViewport((v) => {
-            const center = (v.startMs + v.endMs) / 2;
-            const newDur = (v.endMs - v.startMs) / 0.7;
-            return clampViewport({
-              startMs: center - newDur / 2,
-              endMs: center + newDur / 2,
-            });
-          });
-          break;
-        }
-      }
-    };
-    el.addEventListener("keydown", handleKeyDown);
-    return () => el.removeEventListener("keydown", handleKeyDown);
-  }, [
-    fullDur,
-    selectedSpanId,
-    focusedSpanId,
-    tree.byId,
-    tree.roots,
-    handleResetZoom,
-    handleSpanDoubleClick,
-    onClearSpan,
-    onSelectSpan,
-    clampViewport,
-  ]);
-
   if (spans.length === 0) {
     return (
       <Flex align="center" justify="center" height="full">
@@ -616,266 +293,30 @@ export const FlameView = memo(function FlameView({
     >
       {/* Top bar: breadcrumbs + reset */}
       {(isZoomed || breadcrumbs.length > 0) && (
-        <Flex
-          align="center"
-          justify="space-between"
-          gap={2}
-          paddingX={3}
-          paddingY={1.5}
-          flexShrink={0}
-        >
-          <HStack gap={0.5} flexWrap="nowrap" overflow="hidden" minWidth={0}>
-            <Text
-              as="button"
-              textStyle="xs"
-              color="fg.subtle"
-              cursor="pointer"
-              _hover={{ color: "fg" }}
-              onClick={handleResetZoom}
-              flexShrink={0}
-            >
-              root
-            </Text>
-            {breadcrumbs.map((node, i) => {
-              const isLast = i === breadcrumbs.length - 1;
-              const crumbDur = node.span.endTimeMs - node.span.startTimeMs;
-              const parentDur = node.parent
-                ? node.parent.span.endTimeMs - node.parent.span.startTimeMs
-                : null;
-              const pctOfParent =
-                parentDur !== null && parentDur > 0
-                  ? (crumbDur / parentDur) * 100
-                  : null;
-              return (
-                <HStack key={node.span.spanId} gap={0} minWidth={0}>
-                  <Icon as={LuChevronRight} boxSize={3} color="fg.subtle" />
-                  <HStack
-                    as="button"
-                    gap={1}
-                    paddingX={1}
-                    paddingY={0.5}
-                    borderRadius="sm"
-                    cursor={isLast ? "default" : "pointer"}
-                    _hover={isLast ? undefined : { bg: "bg.muted" }}
-                    onClick={() =>
-                      !isLast && handleSpanDoubleClick(node.span.spanId)
-                    }
-                  >
-                    <Text
-                      textStyle="xs"
-                      color={isLast ? "fg" : "fg.muted"}
-                      fontWeight={isLast ? "medium" : "normal"}
-                      truncate
-                      maxWidth="200px"
-                    >
-                      {node.span.name}
-                    </Text>
-                    <Text textStyle="xs" color="fg.subtle" whiteSpace="nowrap">
-                      {formatDuration(crumbDur)}
-                      {pctOfParent !== null
-                        ? ` · ${formatPercent(pctOfParent)}`
-                        : ""}
-                    </Text>
-                  </HStack>
-                </HStack>
-              );
-            })}
-          </HStack>
-          {isZoomed && (
-            <Tooltip
-              content={
-                <HStack gap={1}>
-                  <Text>Reset zoom</Text>
-                  <Kbd>Esc</Kbd>
-                </HStack>
-              }
-              positioning={{ placement: "top" }}
-            >
-              <Flex
-                as="button"
-                align="center"
-                gap={1}
-                paddingX={2}
-                paddingY={0.5}
-                borderRadius="sm"
-                cursor="pointer"
-                color="fg.muted"
-                _hover={{ bg: "bg.muted", color: "fg" }}
-                onClick={handleResetZoom}
-                flexShrink={0}
-              >
-                <Icon as={LuRotateCcw} boxSize={3} />
-                <Text textStyle="xs">Reset</Text>
-              </Flex>
-            </Tooltip>
-          )}
-        </Flex>
+        <FlameBreadcrumbs
+          breadcrumbs={breadcrumbs}
+          isZoomed={isZoomed}
+          onResetZoom={handleResetZoom}
+          onSpanDoubleClick={handleSpanDoubleClick}
+        />
       )}
 
       {/* Context strip: parent ratio + trace ratio for hovered/focused span */}
-      <Flex
-        align="center"
-        gap={2}
-        paddingX={3}
-        paddingY={1}
-        flexShrink={0}
-        height="26px"
-        borderTopWidth="0.5px"
-        borderBottomWidth="0.5px"
-        borderColor="border.subtle"
-        bg="bg.subtle"
-      >
-        {contextNode && contextInfo ? (
-          <>
-            <Text
-              textStyle="xs"
-              fontWeight="medium"
-              color="fg"
-              truncate
-              maxWidth="220px"
-            >
-              {contextNode.span.name}
-            </Text>
-            <Text textStyle="xs" color="fg.muted" whiteSpace="nowrap">
-              {formatDuration(contextInfo.duration)}
-            </Text>
-            {contextInfo.pctOfParent !== null &&
-              contextInfo.parentName !== null &&
-              contextInfo.parentDuration !== null && (
-                <HStack gap={1} minWidth={0}>
-                  <Text textStyle="xs" color="fg.subtle" whiteSpace="nowrap">
-                    →
-                  </Text>
-                  <Text
-                    textStyle="xs"
-                    color="fg.emphasized"
-                    fontWeight="semibold"
-                    whiteSpace="nowrap"
-                  >
-                    {formatPercent(contextInfo.pctOfParent)}
-                  </Text>
-                  <Text textStyle="xs" color="fg.subtle" whiteSpace="nowrap">
-                    of
-                  </Text>
-                  <Text
-                    textStyle="xs"
-                    color="fg.muted"
-                    truncate
-                    maxWidth="160px"
-                  >
-                    {contextInfo.parentName}
-                  </Text>
-                  <Text textStyle="xs" color="fg.subtle" whiteSpace="nowrap">
-                    ({formatDuration(contextInfo.parentDuration)})
-                  </Text>
-                </HStack>
-              )}
-            {contextInfo.pctOfTrace !== null && (
-              <Text textStyle="xs" color="fg.subtle" whiteSpace="nowrap">
-                · {formatPercent(contextInfo.pctOfTrace)} of trace
-              </Text>
-            )}
-          </>
-        ) : (
-          <Text textStyle="xs" color="fg.subtle">
-            {spans.length} span{spans.length === 1 ? "" : "s"} ·{" "}
-            {formatDuration(fullDur)} · hover a span for details
-          </Text>
-        )}
-      </Flex>
+      <FlameContextStrip
+        contextNode={contextNode}
+        contextInfo={contextInfo}
+        spanCount={spans.length}
+        fullDur={fullDur}
+      />
 
       {/* Time axis: drag to zoom into a range */}
-      <Tooltip
-        content="Drag horizontally to zoom into a range · scroll to zoom · drag flame to pan"
-        positioning={{ placement: "bottom" }}
-        openDelay={400}
-      >
-        <Box
-          ref={timeAxisRef}
-          position="relative"
-          height="28px"
-          flexShrink={0}
-          paddingX={3}
-          cursor="ew-resize"
-          userSelect="none"
-          onPointerDown={handleTimeAxisPointerDown}
-          bg="bg.subtle"
-          _hover={{ bg: "bg.muted" }}
-          transition="background-color 0.1s ease"
-          borderTopWidth="0.5px"
-          borderBottomWidth="0.5px"
-          borderColor="border.subtle"
-          className="flame-time-axis"
-          css={{
-            "&:hover .flame-time-axis-hint": { opacity: 0.95 },
-          }}
-        >
-          {/* Tick lines + labels (ruler-like) */}
-          {ticks.map((tick) => {
-            const offset = (tick.time - viewport.startMs) / dur;
-            if (offset < -0.001 || offset > 1.001) return null;
-            const left = `calc(12px + ${offset} * (100% - 24px))`;
-            return (
-              <Box key={`${tick.label}-${tick.time}`} pointerEvents="none">
-                <Box
-                  position="absolute"
-                  left={left}
-                  bottom={0}
-                  width="1px"
-                  height="6px"
-                  bg="border.emphasized"
-                  opacity={0.6}
-                />
-                <Text
-                  textStyle="xs"
-                  color="fg.muted"
-                  position="absolute"
-                  left={left}
-                  transform="translateX(-50%)"
-                  whiteSpace="nowrap"
-                  userSelect="none"
-                  top="3px"
-                >
-                  {tick.label}
-                </Text>
-              </Box>
-            );
-          })}
-
-          {/* Persistent drag-to-zoom affordance */}
-          <Flex
-            className="flame-time-axis-hint"
-            position="absolute"
-            right={3}
-            top="50%"
-            transform="translateY(-50%)"
-            align="center"
-            gap={1}
-            paddingX={1.5}
-            paddingY={0.5}
-            borderRadius="sm"
-            bg="bg.panel"
-            borderWidth="0.5px"
-            borderColor="border.subtle"
-            color="fg.muted"
-            pointerEvents="none"
-            opacity={0.75}
-            transition="opacity 0.15s ease"
-            boxShadow="xs"
-          >
-            <Text
-              textStyle="2xs"
-              fontWeight="semibold"
-              letterSpacing="0.04em"
-              textTransform="uppercase"
-              whiteSpace="nowrap"
-              lineHeight={1}
-            >
-              ↔ drag to zoom
-            </Text>
-          </Flex>
-        </Box>
-      </Tooltip>
+      <FlameAxis
+        timeAxisRef={timeAxisRef}
+        ticks={ticks}
+        viewport={viewport}
+        dur={dur}
+        onPointerDown={handleTimeAxisPointerDown}
+      />
 
       {/* Flame area — vertical scroll container for the row virtualizer */}
       <Box
