+++ langwatch/src/server/api/routers/__tests__/translate.unit.test.ts
@@ -95,11 +95,10 @@ describe("translateRouter.translate()", () => {
         textToTranslate: "Hola",
       });
 
-      expect(mockGetVercelAIModel).toHaveBeenCalledWith(
+      expect(mockGetVercelAIModel).toHaveBeenCalledWith({
         projectId,
-        undefined,
-        "translate.text",
-      );
+        featureKey: "translate.text",
+      });
     });
 
     it("returns the translated text", async () => {
