+++ langwatch/src/server/app-layer/traces/ai-query.ts
@@ -105,11 +105,10 @@ export async function generateTraceQueryFromPrompt(
   const systemPrompt = buildSystemPrompt(fieldsBlock);
   const messages: ModelMessage[] = [{ role: "user", content: input.prompt }];
 
-  const model = await getVercelAIModel(
-    input.projectId,
-    undefined,
-    "traces.ai_search",
-  );
+  const model = await getVercelAIModel({
+    projectId: input.projectId,
+    featureKey: "traces.ai_search",
+  });
 
   let lastQuery = "";
   let lastError = "Unknown error";
@@ -192,11 +191,10 @@ export async function generateTraceAction(
   const fieldsBlock = await buildFieldsBlock(input);
   const systemPrompt = buildActionSystemPrompt(fieldsBlock);
 
-  const model = await getVercelAIModel(
-    input.projectId,
-    undefined,
-    "traces.ai_search",
-  );
+  const model = await getVercelAIModel({
+    projectId: input.projectId,
+    featureKey: "traces.ai_search",
+  });
 
   let lastError = "Unknown error";
   let lastQuery = "";
