+++ langwatch/src/server/routes/dataset-generate.ts
@@ -82,7 +82,7 @@ ${dataset}`,
     ],
   } as UIMessage);
 
-  const model = await getVercelAIModel(projectId, undefined, "datasets.generator");
+  const model = await getVercelAIModel({ projectId, featureKey: "datasets.generator" });
   const result = streamText({
     model,
     messages: await convertToModelMessages(messages),
